% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_manager.R
\name{format_var}
\alias{format_var}
\title{Get string containing json representation of a single var in the provided environment}
\usage{
format_var(
  envir,
  name,
  abbrev_len = DEFAULT_ABBREV_LEN,
  sort_by = NULL,
  ascending = NULL
)
}
\arguments{
\item{envir}{An environment}

\item{name}{Name of a variable in the environment}

\item{abbrev_len}{The length of elements at which vars should be abbreviated. Pass NULL to prevent abbreviation.}

\item{sort_by}{A vector containing columns to sort (primary first)}

\item{ascending}{A vector of booleans or a single boolean, depending on the data type}
}
\value{
A string with a json representation of a var.
}
\description{
Get string containing json representation of a single var in the provided environment
}
\examples{
x <- c(5,6)
vars_json <- format_var(environment(), "x", NULL)
}
