% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_manager.R
\name{format_vars}
\alias{format_vars}
\title{Get string containing json representation of defined vars in the provided environment}
\usage{
format_vars(envir, abbrev_len = DEFAULT_ABBREV_LEN)
}
\arguments{
\item{envir}{An environment}

\item{abbrev_len}{The length of elements at which vars should be abbreviated. Pass NULL to prevent abbreviation.}
}
\value{
A string with a json list of vars.
}
\description{
Get string containing json representation of defined vars in the provided environment
}
\examples{
vars_json <- format_vars(environment())
}
