% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notification.R
\name{send_notification}
\alias{send_notification}
\title{Send a push notification to the Callisto app running on Mac or iPad}
\usage{
send_notification(title, message)
}
\arguments{
\item{title}{A title for the notification}

\item{message}{A mesage for the notification}
}
\description{
Send a push notification to the Callisto app running on Mac or iPad
}
\examples{
send_notification("Alert", "Your code has finished running")
}
