% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_utils.R
\name{get_var_stats}
\alias{get_var_stats}
\title{Get string containing json statistical summary for each column in a variable}
\usage{
get_var_stats(envir, name, column = NULL)
}
\arguments{
\item{envir}{An environment}

\item{name}{Name of a variable in the environment}

\item{column}{Name of the column to return (returns all columns if not specified)}
}
\value{
A string with a json statistical summary for each column
}
\description{
Get string containing json statistical summary for each column in a variable
}
\examples{
x <- c(5,6)
stats <- get_var_stats(environment(), "x")
}
